Rewrote Avatar - now supports DMs & Discriminators
I rewrote avatar | It may now be used in DMs ( previously yabe would big error )  - Hello I'd like one order of mention, but hold the @ - silent mentions used to autofill the target's name are now supported, as the discrimnator will be stripped. (This may eventually be replaced  in favour of name-autofilling ((@canny)) ) - Added the target's name to the embed's title, so we know whose waifu we're ogling - Modified the help - Added feedback on target not found, a self-destructive message paired with a permanent reaction of disapproval. :zzz: